First wave of changes on Linux

+++ EmuMath/EmuMath/EmuCore/ArithmeticHelpers/CommonMath.h
@@ -202,7 +202,7 @@ namespace EmuCore
 			}
 			else if constexpr(std::is_same_v<FloatingPointIn_, long double>)
 			{
-				return static_cast<FloatingPointOut_>(cosl(static_cast<long double>(val_)));
+				return static_cast<FloatingPointOut_>(cosl(val_));
 			}
 			else if constexpr(EmuCore::TMP::is_static_castable_v<FloatingPointIn_, float>)
 			{
@@ -241,7 +241,7 @@ namespace EmuCore
 			}
 			else if constexpr (std::is_same_v<FloatingPointIn_, long double>)
 			{
-				return static_cast<FloatingPointOut_>(acosl(static_cast<long double>(val_)));
+				return static_cast<FloatingPointOut_>(acosl(val_));
 			}
 			else if constexpr (EmuCore::TMP::is_static_castable_v<FloatingPointIn_, float>)
 			{
@@ -281,7 +281,7 @@ namespace EmuCore
 			}
 			else if constexpr (std::is_same_v<FloatingPointIn_, long double>)
 			{
-				return static_cast<FloatingPointOut_>(sinl(static_cast<long double>(val_)));
+				return static_cast<FloatingPointOut_>(sinl(val_));
 			}
 			else if constexpr (EmuCore::TMP::is_static_castable_v<FloatingPointIn_, float>)
 			{
@@ -320,7 +320,7 @@ namespace EmuCore
 			}
 			else if constexpr (std::is_same_v<FloatingPointIn_, long double>)
 			{
-				return static_cast<FloatingPointOut_>(asinl(static_cast<long double>(val_)));
+				return static_cast<FloatingPointOut_>(asinl(val_));
 			}
 			else if constexpr (EmuCore::TMP::is_static_castable_v<FloatingPointIn_, float>)
 			{
@@ -360,7 +360,7 @@ namespace EmuCore
 			}
 			else if constexpr (std::is_same_v<FloatingPointIn_, long double>)
 			{
-				return static_cast<FloatingPointOut_>(tanl(static_cast<long double>(val_)));
+				return static_cast<FloatingPointOut_>(tanl(val_));
 			}
 			else if constexpr (EmuCore::TMP::is_static_castable_v<FloatingPointIn_, float>)
 			{
@@ -399,7 +399,7 @@ namespace EmuCore
 			}
 			else if constexpr (std::is_same_v<FloatingPointIn_, long double>)
 			{
-				return static_cast<FloatingPointOut_>(atanl(static_cast<long double>(val_)));
+				return static_cast<FloatingPointOut_>(atanl(val_));
 			}
 			else if constexpr (EmuCore::TMP::is_static_castable_v<FloatingPointIn_, float>)
 			{
@@ -505,7 +505,7 @@ namespace EmuCore
 			}
 			else
 			{
-				return static_cast<FloatingPointOut_>(sqrtl(static_cast<long double>(val_)));
+				return static_cast<FloatingPointOut_>(sqrtl(val_));
 			}
 		}
 		else
@@ -529,7 +529,7 @@ namespace EmuCore
 			}
 			else
 			{
-				return static_cast<T_>(floorl(static_cast<long double>(val_)));
+				return static_cast<T_>(floorl(val_));
 			}
 		}
 		else
@@ -560,7 +560,7 @@ namespace EmuCore
 			}
 			else
 			{
-				return static_cast<T_>(ceill(static_cast<long double>(val_)));
+				return static_cast<T_>(ceill(val_));
 			}
 		}
 		else
@@ -591,7 +591,7 @@ namespace EmuCore
 			}
 			else
 			{
-				return static_cast<T_>(truncl(static_cast<long double>(val_)));
+				return static_cast<T_>(truncl(val_));
 			}
 		}
 		else
